Create interface to call TSAN runtime
$ LD_PRELOAD=/usr/lib/gcc/x86_64-linux-gnu/8/libtsan.so ./images/jdk/bin/java -XX:+ThreadSanitizer
works, reports some internal JVM races, and exits with 66, the default exit code when TSAN finds any races.

$ LD_PRELOAD=/usr/lib/gcc/x86_64-linux-gnu/8/libtsan.so TSAN_OPTIONS="exitcode=123" ./images/jdk/bin/java -XX:+ThreadSanitizer ; echo $?
Overriding the default TSAN error exit code works.

With this, the way to provide the TSAN runtime is with LD_PRELOAD. In the future we will have a better way, but this is the current short-term solution so we can iterate faster.