Supports statically linking TSAN runtime into the launcher with Clang
This is a workaround for the version issue with libtsan.so in GCC.
If we build the JDK with Clang, we will have a newer version of the
TSAN runtime statically linked into the launcher:

$ nm images/jdk/bin/java | grep tsan_symbolize
000000000048e640 W __tsan_symbolize_external
000000000048e650 W __tsan_symbolize_external_ex

In the long term, Java TSAN will still work with GCC, but stack trace
symbolization will be broken until OpenJDK can be built with GCC9,
which contains the new version of libtsan.so that we need.

Also need to write distinct values to the global variable in
libAbstractNativeLoop.cpp, otherwise some optimization in Clang
could treat the variable as constant, and TSAN reports "Write of size 1"
instead of "size 4".


Reviewed-by: jcbeyler, aeubanks