Don't reported races on fields marked @LazyInit
This change introduces a new annotation, java.util.concurrent.annotation.LazyInit, which causes TSAN to ignore races on that field.
References are not simply ignored, but a release/acquire is performed on them. This is so that any following accesses to its member variables are not also reported as racy.