[TSAN] Enable field access instrumentation for aarch64
This patch enables field access instrumentation for aarch64.

According to the implementation of x86, this patch inserts TSAN
functions in interpreter (templateTable_aarch64.cpp) at three places
which can cover all field accesses:

  1. getfield_or_static
  2. putfield_or_static
  3. load_field_cp_cache_entry

This patch implements the annotation
'java.util.concurrent.annotation.LazyInit', which causes TSAN to ignore
races on that field. References that marked by @LazyInit are not simply
ignored, but a release/acquire is performed on them. This is so that any
following accesses to its member variables are not also reported as race.

This patch also moves some architecture independent code into common
files.

TODO:
  - Array access instrumentation will be updated in later patch.

Change-Id: I867124a6166c54040ac313fb366861dfc845f18d